requirements.txt
discord.py==2.3.0
facebook_scraper==0.2.59
python-dotenv==1.0.0
